+++ app/src/main/java/com/openclassrooms/rebonnte/domain/Aisle.kt
@@ -4,9 +4,9 @@ import androidx.annotation.Keep
 
 @Keep
 data class Aisle(
-    val aisleId: String,
-    val name: String,
-    val location: String,
-    val description: String,
-    val createdAt: Long
+    val aisleId: String = "",
+    val name: String = "",
+    val location: String = "",
+    val description: String = "",
+    val createdAt: Long = 0
 )
