+++ app/src/main/java/com/openclassrooms/rebonnte/domain/Medicine.kt
@@ -4,11 +4,11 @@ import androidx.annotation.Keep
 
 @Keep
 data class Medicine(
-    val id: String,
-    val name: String,
-    val description: String,
-    val dosage: String,
-    val manufacturer: String,
-    val createdAt: Long
+    val id: String = "",
+    val name: String= "",
+    val description: String = "",
+    val dosage: String = "",
+    val manufacturer: String = "",
+    val createdAt: Long = 0
 
 )
