+++ app/src/main/java/com/openclassrooms/rebonnte/navigation/RebonnteNavHost.kt
@@ -9,6 +9,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.openclassrooms.rebonnte.ui.MainScreen
+import com.openclassrooms.rebonnte.ui.aisle.detail.AisleDetailScreen
 import com.openclassrooms.rebonnte.ui.authentication.LoginScreen
 import com.openclassrooms.rebonnte.ui.authentication.RegisterUserScreen
 import com.openclassrooms.rebonnte.ui.noInternet.NoInternetScreen
@@ -16,7 +17,8 @@ import com.openclassrooms.rebonnte.ui.noInternet.NoInternetScreen
 @Composable
 fun RebonnteNavHost (
     navHostController: NavHostController,
-    startDestination: String
+    startDestination: String,
+    onLogOutAction: ()->Unit
 ) {
     NavHost(
         navController = navHostController,
@@ -24,10 +26,11 @@ fun RebonnteNavHost (
     ) {
 
         //Main screen
-        composable(route = ScreensNav.Main.route){
+        composable(route = ScreensNav.Main.route ){
             //todo main screen conteneur
             MainScreen(
                 onLogOutAction = {
+                    onLogOutAction()
                     navHostController.navigate(ScreensNav.SignIn.route) {
                         //clean the nav backstack
                         popUpTo(0)
@@ -39,10 +42,10 @@ fun RebonnteNavHost (
                     navHostController.navigate(ScreensNav.AddMedicine.route)
                 },
                 onMedicineClicked = {
-                    navHostController.navigate(ScreensNav.MedicineDetails.route)
+                    //navHostController.navigate(ScreensNav.MedicineDetails.createRoute(Medicine.id))
                 },
-                onAisleClicked = {
-                    navHostController.navigate(ScreensNav.AisleDetails.route)
+                onAisleClicked = { aisle ->
+                    navHostController.navigate(ScreensNav.AisleDetails.createRoute(aisle.aisleId))
                 }
 
             )
@@ -60,11 +63,17 @@ fun RebonnteNavHost (
 
         //Aisle Details screen
         composable(route = ScreensNav.AisleDetails.route,
-            arguments = ScreensNav.AisleDetails.navArguments
+            arguments = ScreensNav.AisleDetails.navArguments,
+            enterTransition = {
+                slideInHorizontally(initialOffsetX = { fullWidth -> fullWidth })
+            },
+            exitTransition = {
+                slideOutHorizontally(targetOffsetX = { fullWidth -> fullWidth })
+            }
         ){backStackEntry ->
             val aisleId = backStackEntry.arguments?.getString("aisleId")
                 ?: throw IllegalArgumentException("Aisle ID is required")
-          //  AisleDetailsScreen(aisleId = aisleId))
+            AisleDetailScreen(aisleId = aisleId)
         }
