--- app/src/main/java/com/openclassrooms/rebonnte/ui/aisle/Aisle.kt
@@ -1,3 +1,0 @@
-package com.openclassrooms.rebonnte.ui.aisle
-
-class Aisle(var name: String)
