+++ app/src/main/java/com/openclassrooms/rebonnte/ui/aisle/AisleScreen.kt
@@ -1,45 +1,114 @@
 package com.openclassrooms.rebonnte.ui.aisle
 
-import android.content.Context
-import android.content.Intent
+import android.annotation.SuppressLint
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AccountCircle
 import androidx.compose.material.icons.filled.KeyboardArrowRight
+import androidx.compose.material3.DropdownMenu
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
+import com.openclassrooms.rebonnte.domain.Aisle
 
+@OptIn(ExperimentalMaterial3Api::class)
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 @Composable
 fun AisleScreen(
     viewModel: AisleViewModel = hiltViewModel(),
-    onAisleClicked: (com.openclassrooms.rebonnte.domain.Aisle) -> Unit
+    onAisleClicked: (Aisle) -> Unit,
+    onLogOutAction: () -> Unit
 ) {
     val aisles by viewModel.aisles.collectAsState(initial = emptyList())
     val context = LocalContext.current
 
-    LazyColumn(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        items(aisles) { aisle ->
-            AisleItem(aisle = aisle, onClick = {
-                startDetailActivity(context, aisle.name)
-            })
+    Scaffold(
+        topBar = {
+
+            Column(verticalArrangement = Arrangement.spacedBy((-1).dp)) {
+                TopAppBar(
+                    title = { Text(text = "Aisle")},
+                    actions = {
+                        var expanded by remember { mutableStateOf(false) }
+                        Row(
+                            verticalAlignment = Alignment.CenterVertically,
+                            modifier = Modifier
+                                .padding(end = 8.dp)
+                                .background(MaterialTheme.colorScheme.surface)
+                                .padding(horizontal = 8.dp, vertical = 4.dp)
+                        ) {
+                            Box {
+                                IconButton(onClick = { expanded = true }) {
+                                    Icon(Icons.Default.AccountCircle, contentDescription = null)
+                                }
+                                DropdownMenu(
+                                    expanded = expanded,
+                                    onDismissRequest = { expanded = false },
+                                    offset = DpOffset(x = 0.dp, y = 0.dp)
+                                ) {
+                                    DropdownMenuItem(
+                                        onClick = {
+
+                                            onLogOutAction()
+
+                                            expanded = false
+                                        },
+                                        text = { Text("Log out") }
+                                    )
+                                }
+                            }
+                        }
+                    }
+                )
+            }
+        }
+    ) {paddingValues ->
+        LazyColumn(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(paddingValues)
+        ) {
+            items(aisles, key = { it.aisleId }) { aisle ->
+                AisleItem(aisle = aisle, onClick = {
+                    onAisleClicked(aisle)
+                })
+            }
         }
     }
+
+
+
+
+
 }
 
 @Composable
@@ -51,14 +120,8 @@ fun AisleItem(aisle: Aisle, onClick: () -> Unit) {
             .padding(16.dp),
         horizontalArrangement = Arrangement.SpaceBetween
     ) {
-        Text(text = aisle.name, style = MaterialTheme.typography.bodyMedium)
+        Text(text = aisle.name, style = MaterialTheme.typography.bodyMedium, color = Color.White,modifier = Modifier.align(Alignment.CenterVertically))
         Icon(imageVector = Icons.Default.KeyboardArrowRight, contentDescription = "Arrow")
     }
 }
 
-private fun startDetailActivity(context: Context, name: String) {
-    val intent = Intent(context, AisleDetailActivity::class.java).apply {
-        putExtra("nameAisle", name)
-    }
-    context.startActivity(intent)
-}
