+++ app/src/main/java/com/openclassrooms/rebonnte/ui/medicine/MedicineScreen.kt
@@ -4,61 +4,120 @@ import android.content.Context
 import androidx.compose.runtime.Composable
 
 import android.content.Intent
-import androidx.compose.foundation.clickable
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.KeyboardArrowRight
+import androidx.compose.material.icons.filled.MoreVert
+import androidx.compose.material3.DropdownMenu
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.*
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.DpOffset
 import androidx.hilt.navigation.compose.hiltViewModel
+import com.openclassrooms.rebonnte.domain.Medicine
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun MedicineScreen(
     viewModel: MedicineViewModel = hiltViewModel(),
     onBackClick: () -> Unit,
-    onMedicineClicked: (Medecine) -> Unit
+    onMedicineClicked: (Medicine) -> Unit
     ) {
     val medicines by viewModel.medicines.collectAsState(initial = emptyList())
     val context = LocalContext.current
 
-    LazyColumn(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        items(medicines) { medicine ->
-            MedicineItem(medicine = medicine, onClick = {
-                startDetailActivity(context, medicine.name)
-            })
+    Scaffold(
+        topBar = {
+            var isSearchActive by rememberSaveable { mutableStateOf(false) }
+            var searchQuery by remember { mutableStateOf("") }
+
+            Column(verticalArrangement = Arrangement.spacedBy((-1).dp)) {
+                TopAppBar(
+                    title = { Text(text = "Medicines") },
+                    actions = {
+                        var expanded by remember { mutableStateOf(false) }
+                        Row(
+                            verticalAlignment = Alignment.CenterVertically,
+                            modifier = Modifier
+                                .padding(end = 8.dp)
+                                .background(MaterialTheme.colorScheme.surface)
+                                .padding(horizontal = 8.dp, vertical = 4.dp)
+                        ) {
+                            Box {
+                                IconButton(onClick = { expanded = true }) {
+                                    Icon(Icons.Default.MoreVert, contentDescription = null)
+                                }
+                                DropdownMenu(
+                                    expanded = expanded,
+                                    onDismissRequest = { expanded = false },
+                                    offset = DpOffset(x = 0.dp, y = 0.dp)
+                                ) {
+                                    DropdownMenuItem(
+                                        onClick = {
+                                            //                                 medicineViewModel.sortByNone()
+                                            expanded = false
+                                        },
+                                        text = { Text("Sort by None") }
+                                    )
+                                    DropdownMenuItem(
+                                        onClick = {
+                                            //                                   medicineViewModel.sortByName()
+                                            expanded = false
+                                        },
+                                        text = { Text("Sort by Name") }
+                                    )
+                                    DropdownMenuItem(
+                                        onClick = {
+                                            //                                     medicineViewModel.sortByStock()
+                                            expanded = false
+                                        },
+                                        text = { Text("Sort by Stock") }
+                                    )
+                                }
+                            }
+                        }
+                    }
+                )
+                EmbeddedSearchBar(
+                    query = searchQuery,
+                    onQueryChange = {
+                        //                     medicineViewModel.filterByName(it)
+                        searchQuery = it
+                    },
+                    isSearchActive = isSearchActive,
+                    onActiveChanged = { isSearchActive = it }
+                )
+            }
+
         }
+
+    ){paddingValues ->
+        LazyColumn(
+            modifier = Modifier.fillMaxSize().padding(paddingValues)
+        ) {
+            items(medicines, key = { it.id }) { medicine ->
+                MedicineItem(medicine = medicine, onClick = {
+                    onMedicineClicked(medicine)
+                })
+            }
+        }
+
     }
+
 }
 
-@Composable
-fun MedicineItem(medicine: Medecine, onClick: () -> Unit) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { onClick() }
-            .padding(16.dp),
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        Column {
-            Text(text = medicine.name, style = MaterialTheme.typography.bodyLarge)
-            Text(text = "Stock: ${medicine.stock}", style = MaterialTheme.typography.bodyMedium)
-        }
-        Icon(imageVector = Icons.Default.KeyboardArrowRight, contentDescription = "Arrow")
-    }
-}
 
-private fun startDetailActivity(context: Context, name: String) {
-    val intent = Intent(context, MedicineDetailActivity::class.java).apply {
-        putExtra("nameMedicine", name)
-    }
-    context.startActivity(intent)
-}
+
