+++ app/src/main/java/com/openclassrooms/rebonnte/ui/medicine/MedicineViewModel.kt
@@ -1,7 +1,8 @@
 package com.openclassrooms.rebonnte.ui.medicine
 
 import androidx.lifecycle.ViewModel
-import com.openclassrooms.rebonnte.ui.aisle.Aisle
+import com.openclassrooms.rebonnte.domain.Aisle
+import com.openclassrooms.rebonnte.domain.Medicine
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -11,8 +12,8 @@ import javax.inject.Inject
 
 @HiltViewModel
 class MedicineViewModel @Inject constructor() : ViewModel() {
-    var _medicines = MutableStateFlow<MutableList<Medecine>>(mutableListOf())
-    val medicines: StateFlow<List<Medecine>> get() = _medicines
+    var _medicines = MutableStateFlow<MutableList<Medicine>>(mutableListOf())
+    val medicines: StateFlow<List<Medicine>> get() = _medicines
 
     init {
         _medicines.value = ArrayList() // Initialiser avec une liste vide
@@ -21,20 +22,20 @@ class MedicineViewModel @Inject constructor() : ViewModel() {
     fun addRandomMedicine(aisles: List<Aisle>) {
         val currentMedicines = ArrayList(medicines.value)
         currentMedicines.add(
-            Medecine(
+            Medicine(
                 id = "Medicine " + (currentMedicines.size + 1),
                 name = Random().nextInt(100).toString(),
-                stock = Random().nextInt(100),
-                nameAisle = aisles[Random().nextInt(aisles.size)].name,
-                histories = emptyList()
+              //  stock = Random().nextInt(100),
+            //    nameAisle = aisles[Random().nextInt(aisles.size)].name,
+            //    histories = emptyList()
             )
         )
         _medicines.value = currentMedicines
     }
 
     fun filterByName(name: String) {
-        val currentMedicines: List<Medecine> = medicines.value
-        val filteredMedicines: MutableList<Medecine> = ArrayList()
+        val currentMedicines: List<Medicine> = medicines.value
+        val filteredMedicines: MutableList<Medicine> = ArrayList()
         for (medicine in currentMedicines) {
             if (medicine.name.lowercase(Locale.getDefault())
                     .contains(name.lowercase(Locale.getDefault()))
@@ -51,13 +52,13 @@ class MedicineViewModel @Inject constructor() : ViewModel() {
 
     fun sortByName() {
         val currentMedicines = ArrayList(medicines.value)
-        currentMedicines.sortWith(Comparator.comparing(Medecine::name))
+        currentMedicines.sortWith(Comparator.comparing(Medicine::name))
         _medicines.value = currentMedicines
     }
 
     fun sortByStock() {
         val currentMedicines = ArrayList(medicines.value)
-        currentMedicines.sortWith(Comparator.comparingInt(Medecine::stock))
+       // currentMedicines.sortWith(Comparator.comparingInt(Medicine::stock))
         _medicines.value = currentMedicines
     }
 }
