+++ app/src/main/java/com/openclassrooms/rebonnte/ui/medicine/MedicineWithStockItem.kt
@@ -1,0 +1,35 @@
+package com.openclassrooms.rebonnte.ui.medicine
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.KeyboardArrowRight
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.openclassrooms.rebonnte.domain.MedicineWithStock
+
+@Composable
+fun MedicineWithStockItem(medicine: MedicineWithStock, onClick: (String) -> Unit) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clickable { onClick(medicine.name) }
+            .padding(16.dp),
+        horizontalArrangement = Arrangement.SpaceBetween
+    ) {
+        Column {
+            Text(text = medicine.name, fontWeight = FontWeight.Bold)
+            Text(text = "Stock: ${medicine.quantity}", color = Color.Gray)
+        }
+        Icon(imageVector = Icons.Default.KeyboardArrowRight, contentDescription = "Arrow")
+    }
+}
